+++ tests/Integration/TestHelperTrait.php
@@ -16,6 +16,7 @@ trait TestHelperTrait
 {
     private $first = 'report_1';
     private $second = 'report_2';
+    private $third = 'report_3';
 
     /**
      * @var string[]
@@ -33,6 +34,7 @@ trait TestHelperTrait
             ],
             'totalWarnings' => '21'
         ],
+        'third' => [],
     ];
 
     private function prepareTestData(): void
@@ -49,6 +51,13 @@ trait TestHelperTrait
         $report->setShopId(1);
         $report->setReport($this->data['second']);
         $report->setIssuedAt(new \DateTime('2024-02-01 12:12:12'));
+        $report->save();
+
+        $report = oxNew(Report::class);
+        $report->setId($this->third);
+        $report->setShopId(1);
+        $report->setReport($this->data['third']);
+        $report->setIssuedAt(new \DateTime('2024-03-01 12:12:12'));
         $report->save();
     }
 }